requirements.txt
requests~=2.26.0
mysql-connector-python~=8.0.25
python-dotenv~=0.18.0
discord.py~=1.7.3
discord-py-slash-command~=2.3.2